fix: use unsignedInteger for FK columns to match users.id INT type
users.id is increments() (INT UNSIGNED), not BIGINT. All FK columns
referencing it must be unsignedInteger to satisfy MySQL's FK constraint.

+++ database/migrations/2026_04_19_000002_create_creator_profiles_table.php
@@ -13,8 +13,8 @@ class CreateCreatorProfilesTable extends Migration
         }
 
         Schema::create('creator_profiles', function (Blueprint $table) {
-            $table->unsignedBigInteger('id')->autoIncrement();
-            $table->unsignedBigInteger('user_id');
+            $table->unsignedInteger('id')->autoIncrement();
+            $table->unsignedInteger('user_id');
             $table->string('display_name', 150)->default('');
             $table->text('bio')->nullable();
             $table->string('profile_photo', 255)->nullable();
